starjinx now doesn't free the current meteor if you just leave

+++ NPCs/StarjinxEvent/IStarjinxEnemy.cs
@@ -9,6 +9,7 @@ namespace SpiritMod.NPCs.StarjinxEvent
 	{
 		void DrawPathfinderOutline(SpriteBatch spriteBatch);
 	}
+
 	public static class PathfinderOutlineDraw
 	{
 		public static void DrawAfterImage(SpriteBatch spriteBatch, NPC npc, Rectangle frame, Vector2 offset, Color color, float opacity, float startScale, float endScale, Vector2 origin) => DrawAfterImage(spriteBatch, npc, frame, offset, color, color, opacity, startScale, endScale, origin);
@@ -19,7 +20,7 @@ namespace SpiritMod.NPCs.StarjinxEvent
 			Texture2D tex = Main.npcTexture[npc.type];
 			for (int i = 1; i < 10; i++)
 			{
-				Color color = new Color((int)sbyte.MaxValue, (int)sbyte.MaxValue, (int)sbyte.MaxValue, 0).MultiplyRGBA(Color.Lerp(startColor, endColor, i / 10f)) * 5;
+				Color color = new Color(sbyte.MaxValue, sbyte.MaxValue, sbyte.MaxValue, 0).MultiplyRGBA(Color.Lerp(startColor, endColor, i / 10f)) * 5;
 				spriteBatch.Draw(tex, new Vector2(npc.Center.X, npc.Center.Y) + offset - Main.screenPosition + new Vector2(0, npc.gfxOffY), frame, color * opacity, npc.rotation, origin, MathHelper.Lerp(startScale, endScale, i / 10f), spriteEffects, 0f);
 			}
 		}
